Add hook function for compression_processor of PerMessageCompressionProcessor
PerMessageCompressionProcessor creates its compression processor based
on the request, so we can't pass parameters of compression processor
in advance. Add a hook so that we can pass parameters to the
compression processor.

This CL also contains a bug fix which was detected by the added test.

Review URL: https://codereview.appspot.com/6636047