Use iframe for Chromium performance tests to remove --disable-web-security
Previously:
(a) util_performance_test.js in Blink runs performance tests using
(b) pywebsocket's scripts,
but this required --disable-web-security for cross-origin workers
because (a) and (b) have different origins.

We change this so that:
(a) util_performance_test.js creates
(c) <iframe> with src=pywebsocket's example/*_iframe.html that runs
    the performance tests on the iframe using
(b) pywebsocket's scripts.
Because (b) and (c) have the same origin (but (a) has a different origin),
Tests are run in the same-origin while we have an cross-origin iframe.

This CL adds pywebsocket-side iframe HTML/JavaScript files for (b) and (c).

[1] pywebsocket-side: This CL.
[2] Chromium-side: https://codereview.chromium.org/1521943002/

BUG=https://code.google.com/p/chromium/issues/detail?id=567533